requirements.txt
requests==2.12.4
facebook-sdk
